WL-2295 Fix URL handling of unencoded spaces
Currently URLs are not being encoded when they are first stashed. Since
this is the case it causes an invalid URI when decoding the short URL.
In many cases this will cause a Illegal Argument Exception and interrupt
processing: e.g. this was caught in rewriting html content when copying
a site.

Really we should fix the original encoding problem in the first place, but
this fix for space encodings should fix the lion's share of failures. I
mean WebLearn users only use ascii right? Right?